+++ Changelog.md
@@ -24,5 +24,6 @@
 - User redirected to "/file chat" when trying to use the gui though a terminal.
 - Fixed bug with fileReader not being closed after editing a file.
 - Changed file navigator, text editor & "/plugin" to use more specific permissions.
-- Implemented ability to create new files within the file manager.
+- Implemented ability to create new files.
 - Fixed weird behaviour when interacting with normal items when using the "/file" command.
+- Implemented ability to delete files.